OrchardWatch-ML/growth_model/requirements.txt
Flask-API
flask-cors
logzero